fix WebMonad to provide more flexible typing.
WebMonad was forcing finishWith to accept type "Response" and forcing a 
return value of m Response, which is unnecessary.  Since throwError
of the underlying ErrorT provides and "alternate escape route" you can
let the return value of WebMonad be unconstrained.  This enables you
to create guards that "finishWith" a Response without forcing the entire
monad to contain a value of "Response."  This way guards that need to
end with a Response won't force your entire list of ServerPartTs to
contain a response.