major refactoring of Happstack.Server.HTTP.FileServe. Needs additional testing.
Refactored code. Previously exported functions should have the same
behavior and type signatures as before aside from two exceptions noted below.

The new code is constructed in a far module modular and parameterized
way. This should allow users to implement alternative behavior with
out having to modify the upstream source code or engage in lots of
copy and paste.

Two 'features' were dropped, but could be added on top of the new
interface if they are still valuable.

 1. fileServe no longer returns a fake file if the requested filepath
 starts with TESTH

 2. fileServe no longer supports groups. Previously if a directory
 contained, "foo.c" and "bar.c", then requesting "foo.c,bar.c" would
 cause the server to concat the files together and return them as a
 single file.

I do not believe these features are actively used. I also believe it
should be easy to add they features in a new function built on top of
the new framework.