Fix for inotify event handling.
A previous patch whose description is copied below had a fatal mistake in
inotify event handling. When attempting to register handlers for two files
in the same folder, two event handlers were registered on the same folder.

Due to System.INotify.addWatch semantics, the first registered handler would
be discarded and only the later handler would be called.

The new fix creates a layer on top of inotify, so files are watched by 
observing the folder which contains them. But this time, we register at most
one handler for each folder, and that handler is in charge of dispatching the
events of all observed files in the folder.

> Mon Jul 18 14:57:35 UYT 2011  facundominguez@gmail.com
>
> * Has inotify observe the folder containing a file, rather than the file 
>   itself. Otherwise inotify would stop listening at the file when editing it
>	with vim.
>  
> Vim deletes the file and moves a separate copy of the file onto the original
> location. This caused inotify to stop looking at the file.
>	   
> This patch also has happstack-plugins follow a sourcefile with compile errors
> on the first try to compile it. Previously, if the file had compiling errors
> the first time it was loaded by the application, happstack-plugins would not
> reload it when the file is fixed. This second fix required switching from
> using IORef to using MVars in order to allow IO operations (registering the 
> broken source file) while updating the mapping of the observed files.