Has multiple almost simultaneous file modification events produce only one plugin recompilation.
When two source files are modified almost simultaneously (as a result 
of a cp or a darcs pull command, for instance) the previous behavior was to
compile the plugin twice, that is once for each file modification event.

This patch delays recompilation a few milliseconds after a file modification 
event is received. Thus, if several files are modified close enough in time,
recompilation is delayed until the modifications have settled for a given
timeout period.

When recompilation is in progress and modification events arrive, recompilation
is attempted once more after the current recompilation finishes.