Document and implement the GHC support policy
This is largely a strawman policy, from a short discussion on IRC,
but at least it gives us something concrete to start with.

The idea is that if you have installed the latest HP in the last 18
months, then your GHC is new enough to build the development happs
code.

Currently you need GHC >= 7.0. This might be slightly more restrictive
than the policy above (it's not easy to find out exactly when HP
releases were made, or what versions of GHC they came with), but 7.0
seems a reasonable versino to require at the moment.